pyproject.toml
[tool.poetry]
name = "fhirlight"
version = "0.1.0"
description = ""
authors = ["gecBurton <g.e.c.burton@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
django = "^5.1.4"
djangorestframework = "^3.15.2"
drf-writable-nested = "^0.7.1"
psycopg2-binary = "^2.9.10"
uuid = "^1.30"
django-polymorphic = "^3.1.0"
setuptools = "^75.6.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pytest = "^8.3.4"
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "fhirlight.settings"